Put a cap on the code object size before it is compiled. Default 10,000

+++ src/pyjion/intrins.h
@@ -301,7 +301,7 @@ void PyJit_DecRef(PyObject* value);
 
 int PyJit_PeriodicWork();
 
-PyObject* PyJit_UnicodeJoinArray(PyObject** items, int count);
+PyObject* PyJit_UnicodeJoinArray(PyObject** items, ssize_t count);
 PyObject* PyJit_FormatObject(PyObject* item, PyObject*fmtSpec);
 PyObject* PyJit_FormatValue(PyObject* item);
